feat: Introduce tab completion with host example.
Updated README.md files with information on how to enable tab completion
and made changes to the --host options to highlight the benefit of using
tab completion on an argument with set options.

Made some slight changes to support a new domain variable, in
anticipation of changing URLs, makes it easier to manage.