Feat: Refactor decontamination and upload processes.
Decontamination and upload can now be ran separately, meaning a batch
and it's samples don't need to be ran through hostile before they're
uploaded. Doing this gave the opportunity to do a pretty sizeable
re-factor of the codebase, mainly fleshing out the functionality of the
UploadSample and UploadBatch classes and hpefully making for a cleaner
flow.